+++ trunk/cuisinetunisienne-jpa/src/com/tunisgtug/android/cuisinetunisienne/dao/entity/SortieLine.java
@@ -16,10 +16,10 @@ import com.tunisgtug.android.cuisinetunisienne.dao.entity.Plat;
 public class SortieLine {
 	   
 	@Id
-	private long sorid;
+	private int sorid;
 	
 	@Id
-	private long platid;
+	private int platid;
 	
 	private int quantity;
 	
@@ -40,19 +40,19 @@ public class SortieLine {
 		super();
 	}
 	
-	public long getSorid() {
+	public int getSorid() {
 		return this.sorid;
 	}
 
-	public void setSorid(long sorid) {
+	public void setSorid(int sorid) {
 		this.sorid = sorid;
 	}
 	
-	public long getPlatid() {
+	public int getPlatid() {
 		return this.platid;
 	}
 
-	public void setPlatid(long platid) {
+	public void setPlatid(int platid) {
 		this.platid = platid;
 	}
